Refactors Router / Curriculum
