+++ client/src/app/pages/Landing.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const Landing = (props) => {
+    console.log(props);
+    return <div>Landing Page</div>
+}
+
+export default Landing;
